feat: add functionality to contact form
set the form action url, and the form method, also add "name" attribute
to each form input with their corresponding name. ex: client_name, subject
message, etc.

so when the client sends and email through the form, it
will take the values of each input and send it to the "action" url using
the POST method

Authored-by: Jhonatan Martinez